add logic to check the button states
we're going to use these buttons to know whether or not we should
transition into time setting mode or time display mode. We're also going
to use these buttons to know whether or not to increase or decrease the
time (and by how much) when they're pressed and held within time setting
mode.